feat: Commits old setup
This is an old AstroNvim setup,
there are been modifications to core configs,
and a new user folder has been created where all the new changes should
live.

Do not make further changes to anything other than the user folder.

Still need to investigate why the colour scheme is not working properly,
even though the terminal seems to be displaying the displaying the full
spectrum of colours.