feat(neo-tree): Add Neo Tree to get file info
Neo Tree has a nice way to get information about a file,
by pressing <i> in the file tree, it gives you the name of the file, and
the pathname that you can quickly copy. This is veryy useful.
As such, it's re-enabled, but on another shortcut to only serve that
purpose.